@confused-Techie 2025-09-18
After some review on Discord, it seems this PR does not do what I had hoped. As such I'll close it out and hopefully sometime in the future we can find another way to improve our first load experience